+++ R/sigma_rl.R
@@ -1,0 +1,3 @@
+  sigma_rl <- function(y, sigma_low, rsd_high) {
+    sqrt(sigma_low^2 + y^2 * rsd_high^2)
+  }